[harsh] make it compilable

+++ Src/Harsh/HarshIO.cpp
@@ -68,7 +68,7 @@ void outputInference(const Haplotype& haplotype1,
 
 bool isEmpty(const string& line) {
   int nBit = 0;
-  for (int i = 0; i < line.size(); i++) {
+  for (size_t i = 0; i < line.size(); i++) {
     if (line[i] != '-' && line[i] != ' ' && line[i] != '\t') {
       ++ nBit;
     }
